rename callcenter to reportfixture
