+++ corehq/apps/reportfixtures/tests/__init__.py
@@ -1,5 +1,5 @@
 try:
-    from corehq.apps.callcenter.tests.test_indicator_fixture import *
+    from corehq.apps.reportfixtures.tests.test_indicator_fixture import *
 except ImportError, e:
     # for some reason the test harness squashes these so log them here for clarity
     # otherwise debugging is a pain
